+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+v10.1.97 (2025-03-13)
+-------------------------
+ * Drop CampaignEvent.message field
+
+v10.1.96 (2025-03-13)
+-------------------------
+ * Stop writing CampaignEvent.message
+ * Remove new classifier, make way for llms
+
+v10.1.95 (2025-03-13)
+-------------------------
+ * Fix floweditor grid in chrome
+
 v10.1.94 (2025-03-12)
 -------------------------
  * Read from campaignevent translations instead of message